# backend/internal/httpapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "STA Platform API",
    "version": "v1",
    "description": "Auto-generated shallow spec for the STA Go backend. Method, path, tags and security are accurate; request/response schemas are generic except for a few hand-augmented endpoints. Regenerate with `python3 docs/openapi/generate.py`."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "local"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sta_session"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "adminMFA": {
        "type": "apiKey",
        "in": "header",
        "name": "X-MFA-Code"
      },
      "serviceToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "extraction / agent service token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  },
  "paths": {
...
    "/api/v1/admin/results/batches": {
      "get": {
        "operationId": "get_api_v1_admin_results_batches",
        "summary": "Get admin results batches",
        "tags": [
          "admin-results"
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrf": [],
            "adminMFA": []
          },
          {
            "bearer": [],
            "adminMFA": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/results/batches/{batchID}": {
      "get": {
        "operationId": "get_api_v1_admin_results_batches_batchID",
        "summary": "Get admin results batches batchID",
        "tags": [
          "admin-results"
        ],
        "parameters": [
          {
            "name": "batchID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrf": [],
            "adminMFA": []
          },
          {
            "bearer": [],
            "adminMFA": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/results/import": {
      "post": {
        "operationId": "post_api_v1_admin_results_import",
        "summary": "Create admin results import",
        "tags": [
          "admin-results"
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrf": [],
            "adminMFA": []
          },
          {
            "bearer": [],
            "adminMFA": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/results/{batchID}/inquiries/acceptance-deadline": {
      "post": {
        "operationId": "post_api_v1_admin_results_batchID_inquiries_acceptance_deadline",
        "summary": "Create admin results batchID inquiries acceptance deadline",
        "tags": [
          "admin-results"
        ],
        "parameters": [
          {
            "name": "batchID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrf": [],
            "adminMFA": []
          },
          {
            "bearer": [],
            "adminMFA": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/results/{batchID}/publish": {
      "post": {
        "operationId": "post_api_v1_admin_results_batchID_publish",
        "summary": "Create admin results batchID publish",
        "tags": [
          "admin-results"
        ],
        "parameters": [
          {
            "name": "batchID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrf": [],
            "adminMFA": []
          },
          {
            "bearer": [],
            "adminMFA": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/results/{resultID}/correct": {
      "post": {
        "operationId": "post_api_v1_admin_results_resultID_correct",
        "summary": "Create admin results resultID correct",
        "tags": [
          "admin-results"
        ],
        "parameters": [
          {
            "name": "resultID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrf": [],
            "adminMFA": []
          },
          {
            "bearer": [],
            "adminMFA": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/applications/{applicationID}/candidate-number": {
      "put": {
        "operationId": "put_api_v1_applications_applicationID_candidate_number",
        "summary": "Replace applications applicationID candidate number",
        "tags": [
          "applications"
        ],
        "parameters": [
          {
            "name": "applicationID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{applicationID}/inquiries": {
      "get": {
        "operationId": "get_api_v1_applications_applicationID_inquiries",
        "summary": "Get applications applicationID inquiries",
        "tags": [
          "applications"
        ],
        "parameters": [
          {
            "name": "applicationID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{applicationID}/result": {
      "get": {
        "operationId": "get_api_v1_applications_applicationID_result",
        "summary": "Get applications applicationID result",
        "tags": [
          "applications"
        ],
        "parameters": [
          {
            "name": "applicationID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{applicationID}/willingness": {
      "put": {
        "operationId": "put_api_v1_applications_applicationID_willingness",
        "summary": "Replace applications applicationID willingness",
        "tags": [
          "applications"
        ],
        "parameters": [
          {
            "name": "applicationID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}